UPDATE: implements validation

+++ src/main/java/org/bh_foundation/e_sign/controllers/DocumentController.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.bh_foundation.e_sign.dto.ResponseDto;
 import org.bh_foundation.e_sign.services.data.DocumentService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +16,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
+
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/api/document")
@@ -36,40 +39,65 @@ public class DocumentController {
         return ResponseEntity.ok(documentService.getMine());
     }
 
+    @SuppressWarnings("null")
     @GetMapping("/sign/{id}")
-    public ResponseEntity<ResponseDto<?>> getRequestedById(@PathVariable Long id) {
+    public ResponseEntity<ResponseDto<?>> getRequestedById(@Valid @PathVariable Long id, BindingResult result) {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.ok(documentService.getRequestedById(id));
     }
 
+    @SuppressWarnings("null")
     @GetMapping("/{id}")
     public ResponseEntity<ResponseDto<?>> getMineById(
-            @PathVariable Long id) {
+            @Valid @PathVariable Long id, BindingResult result) {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.ok(documentService.getMineById(id));
     }
 
+    @SuppressWarnings("null")
     @PostMapping("/send")
     public ResponseEntity<ResponseDto<?>> send(
-            @RequestParam String title,
-            @RequestParam boolean order_sign,
-            @RequestParam MultipartFile file,
-            @RequestParam List<Long> signers_id) throws IOException {
+            @Valid @RequestParam String title,
+            @Valid @RequestParam boolean order_sign,
+            @Valid @RequestParam MultipartFile file,
+            @Valid @RequestParam List<Long> signers_id, BindingResult result) throws IOException {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.status(201).body(documentService.send(title, order_sign, file, signers_id));
     }
 
+    @SuppressWarnings("null")
     @PutMapping("/{id}/approve")
-    public ResponseEntity<ResponseDto<?>> approve(@PathVariable Long id) {
+    public ResponseEntity<ResponseDto<?>> approve(@Valid @PathVariable Long id, BindingResult result) {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.ok(documentService.approve(id));
     }
 
+    @SuppressWarnings("null")
     @PutMapping("/{id}/sign")
     public ResponseEntity<ResponseDto<?>> sign(
-            @PathVariable Long id,
-            @RequestParam MultipartFile file) throws IOException {
+            @Valid @PathVariable Long id,
+            @Valid @RequestParam MultipartFile file, BindingResult result) throws IOException {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.ok(documentService.sign(id, file));
     }
 
+    @SuppressWarnings("null")
     @DeleteMapping("/{id}/delete")
-    public ResponseEntity<ResponseDto<?>> delete(@PathVariable Long id) throws IOException {
+    public ResponseEntity<ResponseDto<?>> delete(@Valid @PathVariable Long id, BindingResult result)
+            throws IOException {
+        if (result.hasErrors())
+            return ResponseEntity.status(400)
+                    .body(new ResponseDto<>(400, result.getFieldError().getDefaultMessage(), null));
         return ResponseEntity.status(204).body(documentService.delete(id));
     }
